+++ src/sleep/Makefile
@@ -1,0 +1,1 @@
+../Makefile
